test: CardItem 스토리 작성
